# docs/swagger.yaml
basePath: /
definitions:
  account.ChangePasswordRequest:
    properties:
      new_password:
        example: newpass123
        type: string
      old_password:
        example: currentpass
        type: string
    type: object
  account.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword
        type: string
    type: object
  account.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  account.ProfileResponse:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
      role:
        example: student
        type: string
      uid:
        example: "123456"
        type: string
    type: object
  account.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  account.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      password:
        example: strongpassword
        type: string
    type: object
  account.SetRoleRequest:
    properties:
      role:
        example: teacher
        type: string
      user_id:
        example: 7
        type: integer
    type: object
  account.UpdateProfileRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    type: object
  maps.Building:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  maps.GetShortestPathResponse:
    properties:
      dist:
        type: number
      path:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the TechUp university application.
  title: TechUp API
  version: "1.0"
paths:
  /api/v1/account/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT tokens. Sets access and refresh
        tokens in HTTP-only cookies.
      parameters:
      - description: User login info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - account
  /api/v1/account/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes access and refresh tokens using a valid refresh token
        sent in request body.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New tokens
          schema:
            $ref: '#/definitions/account.LoginResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh JWT tokens
      tags:
      - account
  /api/v1/account/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with role "student"
      parameters:
      - description: User registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - account
  /api/v1/account/secure/change-password:
    post:
      consumes:
      - application/json
      description: Allows a logged-in user to change their password. Authentication
        required via Bearer token.
      parameters:
      - description: Old and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - account
  /api/v1/account/secure/profile:
    get:
      description: Returns current user info based on JWT token passed in Authorization
        header as Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            $ref: '#/definitions/account.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - account
  /api/v1/account/secure/update:
    put:
      consumes:
      - application/json
      description: Allows a logged-in user to update profile information. Authentication
        required via Bearer token.
      parameters:
      - description: Profile update info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated profile
          schema:
            $ref: '#/definitions/account.ProfileResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - account
  /api/v1/account/set-role:
    post:
      consumes:
      - application/json
      description: Allows admin to set the role of another user. Authentication required
        via Bearer token.
      parameters:
      - description: User ID and new role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/account.SetRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set user role
      tags:
      - account
...
  /buildings:
    get:
      description: Returns a list of all buildings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/maps.Building'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all buildings
      tags:
      - map
  /shortest-path:
    get:
      description: Calculates the shortest path between start_room_id and end_room_id
      parameters:
      - description: Start Room ID
        in: query
        name: start_room_id
        required: true
        type: integer
      - description: End Room ID
        in: query
        name: end_room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/maps.GetShortestPathResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get the shortest path between two rooms
      tags:
      - map
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"